code-review-agent-python-bedrock-slack/requirements.txt
aws-cdk-lib==2.208.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.192.0